fix(linux): use uintptr_t for pointer keys in TimeSlicer InstancesMap
Root cause of segfault on 64-bit Linux: DarkSide::ptr2int<int> truncated
64-bit pointers to 32 bits, causing stale/dangling pointers when
int2ptr reconstructed the address in OutputStats.

The InstancesMap stored 'int' keys derived from InstanceInfo* pointers.
On 64-bit systems (LP64), int is 32 bits while pointers are 64 bits.
The upper 32 bits were lost, so int2ptr returned a wrong address.

Fix:
- Change InstancesMap key type from int to uintptr_t
- Update ptr2int<int> to ptr2int<uintptr_t> in AddInstance and Update
- Add defensive null checks in OutputStats for robustness

This affects the HybridServerDispencer TimeSlicer which uses
game_polling_in_separate_thread=1 (config), routing the poller to
worker thread 1 where OutputStats crashed on the first statistics
cycle (~10 seconds after startup).

+++ pw/branches/r1117/Src/Server/AppFramework/TimeSlicer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #include "Interfaces.h"
 #include <System/ThreadHelpers.h>
 #include "System/InlineProfiler.h"
@@ -76,7 +77,7 @@ typedef mpmc_bounded_queue<InstanceInfo*> InstancesPublicQueue;
 typedef nstl::vector<InstancesPublicQueue*> PendingInstances;
 typedef spsc_queue<InstanceInfo*> InstancesQueue;
 typedef nstl::vector<InstanceInfo*> InstancesArray;
-typedef nstl::hash_map<int, int> InstancesMap;
+typedef nstl::hash_map<uintptr_t, int> InstancesMap;
 typedef nstl::map<nstl::string, InstanceStatistics*> InstanceStatisticsMap;
